+++ src/NetscriptFunctions/Flags.ts
@@ -6,7 +6,7 @@ import { NetscriptContext } from "../Netscript/APIWrapper";
 type FlagType = StringConstructor | NumberConstructor | BooleanConstructor | StringConstructor[];
 type FlagsRet = { [key: string]: ScriptArg };
 export function Flags(ctx: NetscriptContext | string[]): (data: unknown) => FlagsRet {
-  const vargs = Array.isArray(ctx) ? ctx : ctx.workerScript.args; 
+  const vargs = Array.isArray(ctx) ? ctx : ctx.workerScript.args;
   return (schema: unknown): FlagsRet => {
     schema = toNative(schema);
     if (!Array.isArray(schema)) throw new Error("flags schema passed in is invalid.");
